+++ react_training/package.json
@@ -12,7 +12,7 @@
     "@testing-library/jest-dom": "5.17.0",
     "@testing-library/react": "13.4.0",
     "@testing-library/user-event": "13.5.0",
-    "axios": "1.6.8",
+    "axios": "^1.6.8",
     "formik": "2.4.5",
     "graphql": "16.8.1",
     "graphql-tag": "2.12.6",
