+++ react_training/src/Day5/Task13/Task13.jsx
@@ -1,57 +1,12 @@
-// Design and implement a Modal component in React. Allow customization of the modal content and provide options to open, close, and toggle the modal. Use this component to display various types of content.
-import React, { useState } from "react";
-import { Box, Button, Typography, Modal } from "../CSS_Files/mui";
-
-const style = {
-  position: "absolute",
-  top: "50%",
-  left: "50%",
-  transform: "translate(-50%, -50%)",
-  width: 400,
-  bgcolor: "background.paper",
-  border: "2px solid #000",
-  boxShadow: 24,
-  p: 4,
-};
+import React from 'react'
+import Model from '../Components/Model'
 
 const Task13 = () => {
-  const [open, setOpen] = useState(false);
-  const handleOpen = () => setOpen(true);
-  const handleClose = () => setOpen(false);
-
   return (
     <>
-      <h2>Modal Component</h2>
-      <div className="modalComponent">
-        <Button
-          style={{ border: "1px solid #000", width: "200px" }}
-          onClick={handleOpen}
-        >
-          Open modal
-        </Button>
-        <Modal
-          open={open}
-          onClose={handleClose}
-          aria-labelledby="modal-modal-title"
-          aria-describedby="modal-modal-description"
-        >
-          <Box sx={style}>
-            <Typography id="modal-modal-title" variant="h6" component="h2">
-              <strong>Toggle the page</strong>
-            </Typography>
-            <Typography id="modal-modal-body" variant="h6" component="h5">
-              Just Click anywhere then the modal component will close
-            </Typography>
-            <Button
-              style={{ border: "1px solid #000", width: "200px" }}
-              onClick={handleClose}
-            >
-              Close modal
-            </Button>
-          </Box>
-        </Modal>
-      </div>
+      <Model a = "Toggle the page" b = "Just Click anywhere then the modal component will close" />
     </>
-  );
-};
-export default Task13;
+  )
+}
+
+export default Task13
