+++ react_training/src/index1.js
@@ -1,0 +1,8 @@
+import Day1 from "./day1/Index";
+import Day2 from "./Day2/Day2";
+import Day3 from "./Day3/Day3";
+import Day4 from "./Day4/Day4";
+import Day5 from "./Day5/Day5";
+import Day6 from "./Day6/Day6";
+
+export { Day1, Day2, Day3, Day4, Day5, Day6 };
